#PASSBOLT-281 #PASSBOLT-270
Override the model and models functions of mad.models.Model which are used by the findOne and findAll functions to automatically create new instances from server responses.
We have to use these functions if we want to benefit to the local cache, which delivers instances from the local cache if they are already present. An instance stays in the cache as long as it is bound.

Add utilities functions toCan, toCakePHP to format data to the respective platform

Add the helper functions (maybe to move in an helper class) getModelReferences, which extract informations from an attribute string. It returns an array of object
{ name : AttributeName, model : reference model, multiplicity : the attribute is multiple or not }