PASSBOLT-825 Cleaning & templates sync & improvements
Check style cleaning
Share configuration between the back end & the front end (app/View/Elements/scriptHeader.ctp).
Remove the try catch in app bootstrap, find a solution later to catch the error, for now it was hard to debug with that.
Add devel tools to watch the current running components, and the sql stack trace, these tools are enabled if the debug variable is not set to 0.
States are now added to the components as classes without js_ prefix.
Make the process of mad.controller.Controller process stronger.
Refactor the breadcrumb presentation.
Refactor the dialog presentation.
Simplify the pwd & ppl workspaces.
Add title attributes to the grid cells.
Change the grid value adapter signature.
Constante APP_URL is now based on cakephp.