Force boolean on tls key value
Cakephp uses this value only as TRUE or FALSE. For whatever reason when the default value is set to 'null' it treats it as TRUE when Docker users either do not pass a variable or are attempting to set it to either NULL or FALSE. It seems to be interpreted as a "non-emptry string which is set" instead of boolean (Docker does not handle quotes which suggests it only passes strings to be interpreted https://docs.docker.com/compose/env-file/#syntax-rules). 

https://github.com/passbolt/passbolt_docker/issues/151 shows setting the key value to null will achieve the desired effect, but this eliminates the ability to use an environment variable.

As tested with https://community.passbolt.com/t/unable-to-disable-tls/4084/14 it appears to honor the false value if forced to boolean.